+++ pkg/snowflake_test.go
@@ -50,18 +50,23 @@ func TestGetConnectionString(t *testing.T) {
 	}
 
 	t.Run("with User/pass", func(t *testing.T) {
-		connectionString := getConnectionString(&config, "password", "")
+		connectionString := getConnectionString(&config, "password", "", "")
 		require.Equal(t, "username:password@account?database=database&role=role&schema=schema&warehouse=warehouse&conf=xxx", connectionString)
 	})
 
 	t.Run("with private key", func(t *testing.T) {
-		connectionString := getConnectionString(&config, "", "privateKey")
+		connectionString := getConnectionString(&config, "", "privateKey", "")
 		require.Equal(t, "username@account?authenticator=SNOWFLAKE_JWT&database=database&privateKey=privateKey&role=role&schema=schema&warehouse=warehouse&conf=xxx", connectionString)
 	})
 
 	t.Run("with User/pass special char", func(t *testing.T) {
-		connectionString := getConnectionString(&config, "p@sswor/d", "")
+		connectionString := getConnectionString(&config, "p@sswor/d", "", "")
 		require.Equal(t, "username:p%40sswor%2Fd@account?database=database&role=role&schema=schema&warehouse=warehouse&conf=xxx", connectionString)
+	})
+
+	t.Run("with token", func(t *testing.T) {
+		connectionString := getConnectionString(&config, "", "", "xxxxxxxx")
+		require.Equal(t, "account?authenticator=oauth&database=database&role=role&schema=schema&token=xxxxxxxx&warehouse=warehouse&conf=xxx", connectionString)
 	})
 
 	config = pluginConfig{
@@ -75,7 +80,7 @@ func TestGetConnectionString(t *testing.T) {
 	}
 
 	t.Run("with string to escape", func(t *testing.T) {
-		connectionString := getConnectionString(&config, "pa$$s&", "")
+		connectionString := getConnectionString(&config, "pa$$s&", "", "")
 		require.Equal(t, "user%40name:pa$$s&@acc@ount?database=dat%40base&role=ro%40le&schema=sch%40ema&warehouse=ware%40house&conf=xxx", connectionString)
 	})
 }