+++ src/types.ts
@@ -28,9 +28,11 @@ export interface SnowflakeOptions extends DataSourceJsonData {
   database?: string;
   schema?: string;
   extraConfig?: string;
+  /** @deprecated use authMethod */
   basicAuth: boolean;
   maxChunkDownloadWorkers?: string;
   customJSONDecoderEnabled: boolean;
+  authMethod: string;
 }
 
 /**
@@ -39,4 +41,5 @@ export interface SnowflakeOptions extends DataSourceJsonData {
 export interface SnowflakeSecureOptions {
   password?: string;
   privateKey?: string;
+  token?: string;
 }